Fix button links on project page
Before, because I was re-using the class name `.projectLink`, the
button links on the project page where stretched over the whole page.

This came from the project links in the project cards on the home page,
which are stretched over the whole "card"/project element.

This needs some more refactoring because it is dirty, but this is a good
fix for now.