Create database directory on every release
To be able to do the initial transfer of the data from Postgres to
SQLite and then replicating of that data we already need the database
directory to exist (so that the database is written into it).

Before, the directory was only created when there is no DATABASE_URL set
and thus SQLite shall be used. As explained above, we need the directory
to exist for the initial trasnfer too.