Use branch names instead of commits for caching
To make sure that the caching is actually beneficial and does not use a random previous
commit's cache, I am splitting caches by branch names.

For a PR, we start with the target branch and save it with the head branch ref. That
means on following pushes, we are restoring from the head.