Use default show toolbar callback in custom one
The point here is that my custom checks are only meant to be added to the existing
checks. I had noticed that in tests the debug toolbar markup is in the page content.
This happened although the test are running with DEBUG=False. Typically having
DEBUG=False should disable the debug toolbar. But, my custom callback did not check for
that. It was not a problem in prod, because the debug toolbar is only installed when
running in debug mode. This is a weird situation, where the debug toolbar is installed,
because the container has the environment variables, but then the tests turn off the
debug mode, but the toolbar still shows...

This is now fixed and my custom show callback uses the default one and prevents the
debug toolbar from showing up in the test in the page markup.