Fix preview when using basic auth
Before, when trying to use preview on a site that has basic auth enabled (e.g. on
staging to avoid indexing) I would keep getting prompted with requests to enter the
basic auth credentials.

It turns out this is caused by the auth information normally getting stripped from the
request headers. Luckily there is a setting to prevent this. Now the the basic auth
credentials work for the other requests too.