Bake default sqlite file into Docker image
This value is likely not changing between the environments and not security relevant. So
I can just bake it into the image. You can still override it if needed, but I don't
intent to. The reason it's a variable is because it is shared between Django settings
and the Litestream settings.