Replace the custom template tag with a decorator
The decorator now adds a method to the dataclass that handles the rendering logic. This
is only done when no rendering method exists yet.

This allows for completely custom rendering logic to be added.

I find, this also reduces the logic and the complexity of the "templex" pattern, because
we can work with fewer moving pieces.